# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onyx AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ProfileInput {
  name?: string;
  avatarInitials?: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  plus: 'plus',
  pro: 'pro',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  avatarInitials?: string;
  avatarUrl?: string | null;
  plan: UserPlan;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Conversation {
  id: number;
  title: string;
  /** @nullable */
  lastMessageAt?: string | null;
  lastMessagePreview?: string | null;
  messageCount: number;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  messages: Message[];
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export type UserSettingsTheme = typeof UserSettingsTheme[keyof typeof UserSettingsTheme];


export const UserSettingsTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserSettings {
  theme: UserSettingsTheme;
  accentColor: string;
  language: string;
  voiceModeEnabled: boolean;
  dataTrainingEnabled: boolean;
}

export type SettingsInputTheme = typeof SettingsInputTheme[keyof typeof SettingsInputTheme];


export const SettingsInputTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface SettingsInput {
  theme?: SettingsInputTheme;
  accentColor?: string;
  language?: string;
  voiceModeEnabled?: boolean;
  dataTrainingEnabled?: boolean;
}

export interface SavedPrompt {
  id: number;
  title: string;
  content: string;
  icon?: string;
  createdAt: string;
}

export interface SavedPromptInput {
  title: string;
  content: string;
  icon?: string;
}

export interface StatsOverview {
  totalConversations: number;
  totalMessages: number;
  savedPrompts: number;
  conversationsToday: number;
}