# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onyx AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from './userPlan';

export interface User {
  id: number;
  email: string;
  name: string;
  avatarInitials?: string;
  avatarUrl?: string | null;
  plan: UserPlan;
  createdAt: Date;
}